requirements.txt
Django==4.1.2
django-admin-ordering==0.13.1
django-ckeditor==6.4.2
django-filter==2.4.0
djangorestframework==3.13.1
nameparser==0.3.16
pyzipcode==2.0.0
us==0.9.1
beautifulsoup4==4.4.1
tzlocal==1.2.2
phonenumbers==8.12.50
six